async: promote runtime to prelude/async.php + soundness rewrite
Spike -> epic. The engine moves out of poc/async/src into a demand-gated prelude
tier (Main.php gates on an `Async` mention and forces Fiber + Io\Poll on), and
every structural defect the spike carried is fixed:

- scope lives on the TASK (Task->owner / Task->scope), not on a scheduler-global
  stack: a spawn can no longer attach to another fiber's group, and a nested
  TaskGroup::run is interleaving-safe;
- REAL cancellation: cancelTask deregisters the victim from the reactor slot /
  timer heap and resumes it with Fiber::throw(CancelledException) at its suspend
  point. A CancelledException is never escalated as a new failure;
- awaitAll is FAIL-FAST: it claims every input BEFORE parking (so a sibling that
  fails while it waits elsewhere is not mistaken for an unhandled failure and
  escalated behind its back -- a *caught* error used to still kill the program at
  exit), cancels + joins the rest, and rethrows the first error;
- awaitAny parks directly on the inputs instead of spawning N watcher fibers, so
  a winner orphans nothing; its AggregateError keys are input-indexed like
  awaitAll's;
- Task::await / joinAll loop on STATE instead of assuming one wake == settled;
  wake() dedupes via Task->queued;
- root scope is joined and the loop reports a DeadlockException instead of
  exiting rc=0 with tasks still parked;
- TaskGroup::run cancels + joins children when the BODY throws (no orphans);
- per-fd waiter slots (readWaiter/writeWaiter) replace the single watcher
  payload, so a reader and a writer on one fd both wake; Write is armed only
  while a writer waits;
- the recv scratch buffer moves onto the Task -- one global buffer that grew via
  free+calloc handed every parked reader a dangling pointer;
- timers become a binary min-heap with lazy delete (was two O(n) scans per tick),
  childSettled is true O(1), the loop snapshots its batch so a self-requeueing
  task cannot starve the reactor, and ioWaiters counts WAITING TASKS not
  registered fds;
- the Scheduler singleton is torn down when run() returns, so repeated async()
  calls in one program get a clean engine.

fiber: injectEx/pendingEx must be \Throwable-typed, not mixed. A mixed property
holds the throwable NaN-boxed, so `throw $cell` reaching a typed
`catch (Foo $e)` dereferenced the boxed word as an object header -> SIGSEGV.
catch (\Throwable) needs no class-id test, which is why fiber_throw never saw it.

poc/async: demos are single files now (no library, no manifest) -- build.sh
compiles them. smoke.php covers the new guarantees: fail-fast + sibling
cancellation, a throwing scope body, and interleaved nested scopes.
