--- poc/async/manticore.json
@@ -1,67 +1,0 @@
-{
-  "applications": [
-    {
-      "name": "async-smoke",
-      "src": "poc/async/src",
-      "output": "poc/async/smoke_bin",
-      "entry": "poc/async/smoke.php",
-      "exclude": []
-    },
-    {
-      "name": "async-echo",
-      "src": "poc/async/src",
-      "output": "poc/async/echo_bin",
-      "entry": "poc/async/echo_server.php",
-      "exclude": []
-    },
-    {
-      "name": "async-load",
-      "src": "poc/async/src",
-      "output": "poc/async/load_bin",
-      "entry": "poc/async/load_client.php",
-      "exclude": []
-    },
-    {
-      "name": "async-http",
-      "src": "poc/async/src",
-      "output": "poc/async/http_bin",
-      "entry": "poc/async/http_server.php",
-      "exclude": []
-    },
-    {
-      "name": "async-chan",
-      "src": "poc/async/src",
-      "output": "poc/async/chan_bin",
-      "entry": "poc/async/chan_demo.php",
-      "exclude": []
-    },
-    {
-      "name": "async-http-transparent",
-      "src": "poc/async/src",
-      "output": "poc/async/http_transparent_bin",
-      "entry": "poc/async/http_transparent.php",
-      "exclude": []
-    },
-    {
-      "name": "async-capture",
-      "src": "poc/async/src",
-      "output": "poc/async/capture_bin",
-      "entry": "poc/async/capture.php",
-      "exclude": []
-    },
-    {
-      "name": "async-spawn",
-      "src": "poc/async/src",
-      "output": "poc/async/spawn_bin",
-      "entry": "poc/async/spawncost.php",
-      "exclude": []
-    },
-    {
-      "name": "async-io",
-      "src": "poc/async/src",
-      "output": "poc/async/async_io",
-      "entry": "poc/async/async-io.php",
-      "exclude": []
-    }
-  ]
-}
